requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
pillow
python-multipart
aiofiles